🔨 Refactor: 문서 구조 수정 및 코드 컨벤션 적용
