+++ src/app/layout.tsx
@@ -3,9 +3,9 @@ import "./globals.css";
 import Script from "next/script";
 import { Toaster } from "@/src/components/ui/sonner";
 import RootProvider from "../contexts/RootProvider";
-import FCMNotification from "../components/fcm/FCMNotification";
+import FCMNotification from "../hooks/use-fcm-notification";
 import { Providers } from "../components/display-mode/ThemeProvider";
-import ThemeColorUpdater from "../components/home/ThemeColorUpdater";
+import ThemeColorUpdater from "../hooks/use-theme-color-updater";
 import { getMetadata } from "../utils";
 export const metadata: Metadata = getMetadata();
