+++ src/features/feedback/feedback-common/components/feedback-dropdown-menu.tsx
@@ -11,7 +11,7 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "@/src/components/ui/dropdown-menu";
-import { ArrowDownIcon } from "../../utils/icons";
+import { ArrowDownIcon } from "../../../../components/utils/icons";
 
 export default function DropDownMenu({
   addCategory,
