+++ src/features/feedback/feedback-mobile/components/feedback-mobile-toolbar.tsx
@@ -3,8 +3,8 @@
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
 import React, { useState, useEffect, useMemo } from "react";
 import useToolbarStore from "@/src/store/feedback/useToolbarStore";
-import FeedbackRadioButton from "./FeedbackRadioButton";
-import FeedbackSearchbar from "./FeedbackSearchbar";
+import FeedbackRadioButton from "../../feedback-common/components/feedback-radio-button";
+import FeedbackSearchbar from "../../feedback-common/components/feedback-searchbar";
 
 export default function MobileToolbar() {
   const { isMobileToolbarOpen, closeMobileToolbar } = useToolbarStore();
