+++ src/features/home/home-screen/components/home-bg.tsx
